+++ backend/app/webserver/controllers/project/get-projects-category-controller.js
@@ -34,8 +34,8 @@ async function getProjectsFilter(req, res, next) {
 
   let connection;
   try {
-    const sqlQuery = `SELECT t.id, t.title, t.subtitle, t.created_at, t.updated_at, t.category, t.text, t.ubication, u.name, u.first_name, u.id AS user
-      FROM project t JOIN user u ON t.user_id = u.id ORDER BY t.created_at DESC`;
+    const sqlQuery = `SELECT t.id, t.title, t.subtitle, t.ubication, t.text, t.created_at, t.updated_at, t.category, u.name, u.first_name, u.id AS user
+    FROM project t JOIN user u ON t.user_id = u.id WHERE category = ? ORDER BY t.created_at DESC`;
     connection = await mysqlPool.getConnection();
     const [rows] = await connection.execute(sqlQuery, [categoryCap]);
     connection.release();
