+++ frontend/src/components/Navbar.js
@@ -175,103 +175,7 @@ class App extends React.Component {
                     Tecnología
                   </a>
                 </button>
-                <hr className="style1" />
-                <button
-                  onClick={() => this.changeStyle('forum')}
-                  className="forum w-full px-4 py-2 text-white hover:bg-orange-800"
-                >
-                  Foro ▼
-                  <a
-                    href="/"
-                    className="hidden relative identifier block px-4 py-2 text-center text-white hover:text-black "
-                  >
-                    Todos los temas
-                  </a>
-                  <a
-                    href="/forums/arte"
-                    className="hidden relative identifier block px-4 py-2 text-center text-white hover:text-black "
-                  >
-                    Arte
-                  </a>
-                  <a
-                    href="/forums/artesanías"
-                    className="hidden relative identifier block px-4 py-2 text-center text-white hover:text-black "
-                  >
-                    Artesanías
-                  </a>
-                  <a
-                    href="/forums/cine"
-                    className="hidden relative identifier block px-4 py-2 text-center text-white hover:text-black "
-                  >
-                    Cine
-                  </a>
-                  <a
-                    href="/forums/cómics"
-                    className="hidden relative identifier block px-4 py-2 text-center text-white hover:text-black "
-                  >
-                    Cómics
-                  </a>
-                  <a
-                    href="/forums/danza"
-                    className="hidden relative identifier block px-4 py-2 text-center text-white hover:text-black "
-                  >
-                    Danza
-                  </a>
-                  <a
-                    href="/forums/diseño"
-                    className="hidden relative identifier block px-4 py-2 text-center text-white hover:text-black "
-                  >
-                    Diseño
-                  </a>
-                  <a
-                    href="/forums/fotografía"
-                    className="hidden relative identifier block px-4 py-2 text-center text-white hover:text-black "
-                  >
-                    Fotografía
-                  </a>
-                  <a
-                    href="/forums/juegos"
-                    className="hidden relative identifier block px-4 py-2 text-center text-white hover:text-black "
-                  >
-                    Juegos
-                  </a>
-                  <a
-                    href="/forums/moda"
-                    className="hidden relative identifier block px-4 py-2 text-center text-white hover:text-black "
-                  >
-                    Moda
-                  </a>
-                  <a
-                    href="/forums/música"
-                    className="hidden relative identifier block px-4 py-2 text-center text-white hover:text-black "
-                  >
-                    Música
-                  </a>
-                  <a
-                    href="/forums/periodismo"
-                    className="hidden relative identifier block px-4 py-2 text-center text-white hover:text-black "
-                  >
-                    Periodismo
-                  </a>
-                  <a
-                    href="/forums/publicaciones"
-                    className="hidden relative identifier block px-4 py-2 text-center text-white hover:text-black "
-                  >
-                    Publicaciones
-                  </a>
-                  <a
-                    href="/forums/teatro"
-                    className="hidden relative identifier block px-4 py-2 text-center text-white hover:text-black "
-                  >
-                    Teatro
-                  </a>
-                  <a
-                    href="/forums/tecnología"
-                    className="hidden relative identifier block px-4 py-2 text-center text-white hover:text-black "
-                  >
-                    Tecnología
-                  </a>
-                </button>
+
                 <hr className="style1" />
                 <button
                   onClick={() => this.changeStyle('configuration')}
